Implemented hover and jump sound effects
There great. There's 2 things to note though. 1) For the jump, the startup of the sound makes it feel like the jumpsfx is a little delayed, so I think we should cut out a bit of the beginning so it can feel a little more instantaneous. 2) The hover sound is nice, but when I stop the hover sound, it stops instantly and sounds a bit too robotic. It would probably be best to add a fade out event. Should be quick. Both of these things are done in fmod studio, so I'll talk to luke about this.