Finished build 2.0
This is kinda what I wanted to end with by the end of the game jam. It's not perfect, and there's space for a lot more, but considering the game jam was pretty small, I thought this was a decent place to stop.
Differences from this build to the one at the end of the game jam are:
Bloom effect on snow (snow glows), better controls, another level, a title screen, small bug fixes, patch for luke's level (it was a bit too hard, so I made a bit easier and more intuitive), and particle effects instead of green boxes for checkpoints. It was a nice experience and I learned a lot :)